added e2e tests for create, update applicant
